+++ packages/gatsby/cache-dir/production-app.js
@@ -75,7 +75,7 @@ apiRunnerAsync(`onClientEntry`).then(() => {
                   <RouteHandler
                     path={encodeURI(
                       pageResources.page.path === `/404.html`
-                        ? location.pathname
+                        ? stripPrefix(location.pathname, __BASE_PATH__)
                         : pageResources.page.matchPath ||
                             pageResources.page.path
                     )}
@@ -106,9 +106,7 @@ apiRunnerAsync(`onClientEntry`).then(() => {
     pagePath &&
     __BASE_PATH__ + pagePath !== browserLoc.pathname &&
     !(
-      loader.pathFinder.findMatchPath(
-        stripPrefix(browserLoc.pathname, __BASE_PATH__)
-      ) ||
+      loader.findMatchPath(stripPrefix(browserLoc.pathname, __BASE_PATH__)) ||
       pagePath === `/404.html` ||
       pagePath.match(/^\/404\/?$/) ||
       pagePath.match(/^\/offline-plugin-app-shell-fallback\/?$/)