+++ www/gatsby-node.js
@@ -47,6 +47,12 @@ exports.createPages = ({ graphql, actions, reporter }) => {
   const { createPage, createRedirect } = actions
 
   createRedirect({
+    fromPath: `/docs/themes/api-reference`,
+    toPath: `/docs/theme-api/`,
+    isPermanent: true,
+  })
+
+  createRedirect({
     fromPath: `/docs/component-css/`, // Merged Component CSS and CSS Modules
     toPath: `/docs/css-modules/`,
     isPermanent: true,
@@ -282,7 +288,12 @@ exports.createPages = ({ graphql, actions, reporter }) => {
   })
   createRedirect({
     fromPath: `/docs/image-tutorial/`,
-    toPath: `/tutorial/image-tutorial/`,
+    toPath: `/tutorial/wordpress-image-tutorial/`,
+    isPermanent: true,
+  })
+  createRedirect({
+    fromPath: `/tutorial/image-tutorial/`,
+    toPath: `/tutorial/wordpress-image-tutorial/`,
     isPermanent: true,
   })
   createRedirect({
@@ -936,6 +947,28 @@ exports.sourceNodes = ({ actions: { createTypes }, schema }) => {
     type Airtable implements Node {
       id: ID!
       data: AirtableData
+    }
+
+    type SitesYaml implements Node {
+      title: String!
+      main_url: String!
+      url: String!
+      source_url: String
+      featured: Boolean
+      categories: [String]!
+      built_by: String
+      built_by_url: String
+      description: String
+      screenshotFile: Screenshot # added by gatsby-transformer-screenshot
+    }
+
+    type StartersYaml implements Node {
+      url: String!
+      repo: String!
+      description: String
+      tags: [String!]
+      features: [String!]
+      screenshotFile: Screenshot # added by gatsby-transformer-screenshot
     }
 
     type AirtableData @dontInfer {