+++ packages/gatsby-source-drupal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.2.7...gatsby-source-drupal@3.2.8) (2019-07-22)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.2.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.2.6...gatsby-source-drupal@3.2.7) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
 ## [3.2.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.2.5...gatsby-source-drupal@3.2.6) (2019-07-12)
 
 ### Bug Fixes